Replace old PerMessageDeflateExtension with new implementation
Instead of adding WebSocketPerMessageDeflateExtension side by side with
the old PerMessageDeflateExtension, rename the new implementation to the
old name to replace it.
